[Improve][Connector-V2][ioTdb] Refactor iceberg connector e2e test cases
