+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-iceberg-e2e/pom.xml
@@ -17,56 +17,26 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>seatunnel-e2e</artifactId>
         <groupId>org.apache.seatunnel</groupId>
+        <artifactId>seatunnel-connector-v2-e2e</artifactId>
         <version>${revision}</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-    <packaging>pom</packaging>
-    <modules>
-        <module>connector-assert-e2e</module>
-        <module>connector-jdbc-e2e</module>
-        <module>connector-redis-e2e</module>
-        <module>connector-cdc-sqlserver-e2e</module>
-        <module>connector-clickhouse-e2e</module>
-        <module>connector-starrocks-e2e</module>
-        <module>connector-influxdb-e2e</module>
-        <module>connector-amazondynamodb-e2e</module>
-        <module>connector-file-local-e2e</module>
-        <module>connector-cassandra-e2e</module>
-        <module>connector-neo4j-e2e</module>
-        <module>connector-http-e2e</module>
-        <module>connector-rabbitmq-e2e</module>
-        <module>connector-kafka-e2e</module>
-        <module>connector-doris-e2e</module>
-        <module>connector-fake-e2e</module>
-        <module>connector-elasticsearch-e2e</module>
-        <module>connector-iotdb-e2e</module>
-        <module>connector-cdc-mysql-e2e</module>
-        <module>connector-iceberg-e2e</module>
-        <module>connector-iceberg-hadoop3-e2e</module>
-    </modules>
 
-    <artifactId>seatunnel-connector-v2-e2e</artifactId>
+    <artifactId>connector-iceberg-e2e</artifactId>
 
     <dependencies>
         <dependency>
             <groupId>org.apache.seatunnel</groupId>
-            <artifactId>seatunnel-e2e-common</artifactId>
-            <version>${project.version}</version>
-            <type>test-jar</type>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.seatunnel</groupId>
-            <artifactId>seatunnel-flink-starter</artifactId>
+            <artifactId>connector-iceberg</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.apache.seatunnel</groupId>
-            <artifactId>seatunnel-spark-starter</artifactId>
+            <artifactId>seatunnel-hadoop3-3.1.4-uber</artifactId>
             <version>${project.version}</version>
+            <classifier>optional</classifier>
             <scope>test</scope>
         </dependency>
     </dependencies>
