+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-iceberg-e2e/src/test/resources/iceberg/iceberg_source.conf
@@ -1,0 +1,84 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+######
+###### This config file is a demonstration of streaming processing in seatunnel config
+######
+
+env {
+  execution.parallelism = 2
+  job.mode = "BATCH"
+  execution.checkpoint.interval = 10
+}
+
+source {
+  Iceberg {
+    fields {
+      f2 = "boolean"
+      f1 = "bigint"
+      f3 = "int"
+      f4 = "bigint"
+      f5 = "float"
+      f6 = "double"
+      f7 = "date"
+      f9 = "timestamp"
+      f10 = "timestamp"
+      f11 = "string"
+      f12 = "bytes"
+      f13 = "bytes"
+      f14 = "decimal(19,9)"
+      f15 = "array<int>"
+      f16 = "map<string, int>"
+    }
+    catalog_name = "seatunnel"
+    catalog_type = "hadoop"
+    warehouse = "file:///tmp/seatunnel/iceberg/flink/"
+    namespace = "database1"
+    table = "source"
+  }
+}
+
+transform {
+}
+
+sink {
+  Console {
+  }
+  Assert {
+    rules =
+      {
+        field_rules = [
+          {
+            field_name = f1
+            field_type = long
+            field_value = [
+              {
+                rule_type = NOT_NULL
+              },
+              {
+                rule_type = MIN
+                rule_value = 0
+              },
+              {
+                rule_type = MAX
+                rule_value = 99
+              }
+            ]
+          }
+        ]
+      }
+  }
+}
