+++ seatunnel-e2e/seatunnel-flink-connector-v2-e2e/pom.xml
@@ -31,8 +31,6 @@
         <module>connector-jdbc-flink-e2e</module>
         <module>connector-datahub-flink-e2e</module>
         <module>connector-mongodb-flink-e2e</module>
-        <module>connector-iceberg-flink-e2e</module>
-        <module>connector-iceberg-hadoop3-flink-e2e</module>
     </modules>
 
     <dependencies>
